WS6: regenerate dartpy stubs for the ECS facade; drop simulation_experimental.pyi
- Regenerate python/stubs after the World name-swap: simulation.pyi is now the
  ECS facade shape; __init__.pyi re-exports the ECS World flat; gui stub gains
  RenderWorld (+ extract_renderables/run_application typed against it).
- Delete the orphaned python/stubs/dartpy/simulation_experimental.pyi (the
  module no longer exists).
- Define dartpy.gui.world_render_frame() as a real function in
  _world_render_bridge.py (attached in __init__) so the stub generator emits a
  correct `from dartpy._world_render_bridge import world_render_frame` instead of
  the bogus `from dartpy.dynamics import World as world_render_frame` produced
  when it was a bound staticmethod.

Verified: stubs regenerate cleanly; dartpy.gui.world_render_frame() returns a
Frame; bridge + render SimpleFrame work under -W error.
