+++ python/stubs/dartpy/__init__.pyi
@@ -12,7 +12,6 @@ from . import io
 from . import math
 from . import optimizer
 from . import simulation
-from . import simulation_experimental
 from . import utils
 
 from .collision import (
@@ -256,19 +255,73 @@ from .optimizer import (
 )
 
 from .simulation import (
-    Bullet,
-    CollisionDetectorType,
-    ConstraintSolver,
-    Dantzig,
-    Dart,
-    Experimental,
-    Fcl,
-    LcpSolverType,
-    Lemke,
-    Ode,
-    Pgs,
+    ActuatorType,
+    ClosureDynamicsPolicy,
+    ClosureKinematicsPolicy,
+    CollisionBody,
+    CollisionQueryOptions,
+    CollisionShape,
+    CollisionShapeType,
+    Contact,
+    ContactGradientMode,
+    ContactSolverMethod,
+    DeformableBody,
+    DeformableBodyOptions,
+    DeformableDirichletBoundaryCondition,
+    DeformableEdge,
+    DeformableMaterialProperties,
+    DeformableNeumannBoundaryCondition,
+    DeformableSceneBodyInfo,
+    DeformableSceneDiagnostics,
+    DeformableSceneInfo,
+    DeformableSceneLoadOptions,
+    DeformableSolverDiagnostics,
+    DeformableSolverOptions,
+    DeformableSurfaceTriangle,
+    DeformableTetrahedron,
+    FixedFrame,
+    Frame,
+    FreeFrame,
+    Joint,
+    JointSpec,
+    JointType,
+    Link,
+    LoopClosure,
+    LoopClosureFamily,
+    LoopClosureResidual,
+    LoopClosureResidualCoordinates,
+    LoopClosureRuntimePolicy,
+    LoopClosureSpec,
+    ModelFormat,
+    Multibody,
+    MultibodyOptions,
+    PhysicalParameter,
+    ReadOptions,
+    RigidBody,
+    RigidBodyOptions,
+    RigidBodySolver,
+    RootJointType,
+    SkeletonLoadOptions,
+    SkeletonToMultibodyOptions,
+    StateSpace,
+    StateVariable,
+    StepDerivatives,
+    StepGradient,
     World,
-    WorldConfig,
+    WorldSyncStage,
+    add_skeleton,
+    add_world,
+    build_multibodies_from_world,
+    build_multibody_from_skeleton,
+    collect_deformable_scene_diagnostics,
+    is_accelerated_deformable_solve_available,
+    is_accelerated_deformable_solve_enabled,
+    load_deformable_scene,
+    load_gmsh_tet_mesh,
+    load_obj_triangle_mesh,
+    load_point_set,
+    load_seg_line_mesh,
+    set_accelerated_deformable_solve,
 )
 
 __all__: list[str] = [
@@ -280,32 +333,50 @@ __all__: list[str] = [
     "BodyNode",
     "BodyNodeCollisionFilter",
     "BoxShape",
-    "Bullet",
     "CapsuleShape",
     "Chain",
     "ChainCriteria",
+    "ClosureDynamicsPolicy",
+    "ClosureKinematicsPolicy",
     "CollisionAspect",
+    "CollisionBody",
     "CollisionDetector",
-    "CollisionDetectorType",
     "CollisionFilter",
     "CollisionGroup",
     "CollisionObject",
     "CollisionOption",
+    "CollisionQueryOptions",
     "CollisionResult",
+    "CollisionShape",
+    "CollisionShapeType",
     "Composite",
     "ConeShape",
     "ConstraintBase",
-    "ConstraintSolver",
+    "Contact",
+    "ContactGradientMode",
+    "ContactSolverMethod",
     "ContinuousCollisionAdvancement",
     "ContinuousCollisionHit",
     "ContinuousCollisionOption",
     "ContinuousCollisionResult",
     "Coupler",
     "CylinderShape",
-    "Dantzig",
-    "Dart",
     "DartCollisionDetector",
     "DefaultActuatorType",
+    "DeformableBody",
+    "DeformableBodyOptions",
+    "DeformableDirichletBoundaryCondition",
+    "DeformableEdge",
+    "DeformableMaterialProperties",
+    "DeformableNeumannBoundaryCondition",
+    "DeformableSceneBodyInfo",
+    "DeformableSceneDiagnostics",
+    "DeformableSceneInfo",
+    "DeformableSceneLoadOptions",
+    "DeformableSolverDiagnostics",
+    "DeformableSolverOptions",
+    "DeformableSurfaceTriangle",
+    "DeformableTetrahedron",
     "DegreeOfFreedom",
     "Detachable",
     "DynamicJointConstraint",
@@ -314,10 +385,10 @@ __all__: list[str] = [
     "EndEffector",
     "Entity",
     "EulerJoint",
-    "Experimental",
     "FORCE",
-    "Fcl",
+    "FixedFrame",
     "Frame",
+    "FreeFrame",
     "FreeJoint",
     "FreeJointProperties",
     "Function",
@@ -335,25 +406,34 @@ __all__: list[str] = [
     "Joint",
     "JointConstraint",
     "JointCoulombFrictionConstraint",
+    "JointSpec",
+    "JointType",
     "LOCKED",
-    "LcpSolverType",
-    "Lemke",
     "LineSegmentShape",
+    "Link",
     "Linkage",
     "LinkageCriteria",
     "LocalResource",
     "LocalResourceRetriever",
+    "LoopClosure",
+    "LoopClosureFamily",
+    "LoopClosureResidual",
+    "LoopClosureResidualCoordinates",
+    "LoopClosureRuntimePolicy",
+    "LoopClosureSpec",
     "MIMIC",
     "MeshShape",
     "MetaSkeleton",
     "MimicConstraintType",
     "MimicDofProperties",
+    "ModelFormat",
     "Motor",
+    "Multibody",
+    "MultibodyOptions",
     "NullFunction",
     "Observer",
-    "Ode",
     "PASSIVE",
-    "Pgs",
+    "PhysicalParameter",
     "PlanarJoint",
     "PlaneShape",
     "PointCloudShape",
@@ -366,12 +446,17 @@ __all__: list[str] = [
     "RayHit",
     "RaycastOption",
     "RaycastResult",
+    "ReadOptions",
     "Resource",
     "ResourceRetriever",
     "ResourceSeekType",
     "RevoluteJoint",
     "RevoluteJointConstraint",
     "RevoluteJointProperties",
+    "RigidBody",
+    "RigidBodyOptions",
+    "RigidBodySolver",
+    "RootJointType",
     "SE3",
     "SE3Tangent",
     "SERVO",
@@ -384,9 +469,15 @@ __all__: list[str] = [
     "ShapeNode",
     "SimpleFrame",
     "Skeleton",
+    "SkeletonLoadOptions",
+    "SkeletonToMultibodyOptions",
     "Solver",
     "SolverProperties",
     "SphereShape",
+    "StateSpace",
+    "StateVariable",
+    "StepDerivatives",
+    "StepGradient",
     "Stopwatch",
     "Subject",
     "Support",
@@ -404,9 +495,14 @@ __all__: list[str] = [
     "WeldJointConstraint",
     "WholeBodyIK",
     "World",
-    "WorldConfig",
+    "WorldSyncStage",
     "ZeroDofJoint",
     "ZeroDofJointProperties",
+    "add_skeleton",
+    "add_world",
+    "build_multibodies_from_world",
+    "build_multibody_from_skeleton",
+    "collect_deformable_scene_diagnostics",
     "collision",
     "common",
     "constraint",
@@ -459,6 +555,13 @@ __all__: list[str] = [
     "inf",
     "info",
     "io",
+    "is_accelerated_deformable_solve_available",
+    "is_accelerated_deformable_solve_enabled",
+    "load_deformable_scene",
+    "load_gmsh_tet_mesh",
+    "load_obj_triangle_mesh",
+    "load_point_set",
+    "load_seg_line_mesh",
     "log",
     "math",
     "matrixToEulerXYX",
@@ -488,8 +591,8 @@ __all__: list[str] = [
     "rand_se3_tangent",
     "rand_so3",
     "rand_so3_tangent",
+    "set_accelerated_deformable_solve",
     "simulation",
-    "simulation_experimental",
     "split",
     "tic",
     "toLower",
