--- python/stubs/dartpy/simulation_experimental.pyi
@@ -1,1873 +1,0 @@
-from __future__ import annotations
-
-__all__: list[str] = [
-    "ActuatorType",
-    "ClosureDynamicsPolicy",
-    "ClosureKinematicsPolicy",
-    "CollisionBody",
-    "CollisionQueryOptions",
-    "CollisionShape",
-    "CollisionShapeType",
-    "Contact",
-    "ContactGradientMode",
-    "ContactSolverMethod",
-    "DeformableBody",
-    "DeformableBodyOptions",
-    "DeformableDirichletBoundaryCondition",
-    "DeformableEdge",
-    "DeformableMaterialProperties",
-    "DeformableNeumannBoundaryCondition",
-    "DeformableSceneBodyInfo",
-    "DeformableSceneDiagnostics",
-    "DeformableSceneInfo",
-    "DeformableSceneLoadOptions",
-    "DeformableSolverDiagnostics",
-    "DeformableSolverOptions",
-    "DeformableSurfaceTriangle",
-    "DeformableTetrahedron",
-    "FixedFrame",
-    "Frame",
-    "FreeFrame",
-    "Joint",
-    "JointSpec",
-    "JointType",
-    "Link",
-    "LoopClosure",
-    "LoopClosureFamily",
-    "LoopClosureResidual",
-    "LoopClosureResidualCoordinates",
-    "LoopClosureRuntimePolicy",
-    "LoopClosureSpec",
-    "ModelFormat",
-    "Multibody",
-    "MultibodyOptions",
-    "PhysicalParameter",
-    "ReadOptions",
-    "RigidBody",
-    "RigidBodyOptions",
-    "RigidBodySolver",
-    "RootJointType",
-    "SkeletonLoadOptions",
-    "SkeletonToMultibodyOptions",
-    "StateSpace",
-    "StateVariable",
-    "StepDerivatives",
-    "StepGradient",
-    "World",
-    "WorldSyncStage",
-    "add_skeleton",
-    "add_world",
-    "build_multibodies_from_world",
-    "build_multibody_from_skeleton",
-    "collect_deformable_scene_diagnostics",
-    "is_accelerated_deformable_solve_available",
-    "is_accelerated_deformable_solve_enabled",
-    "load_deformable_scene",
-    "load_gmsh_tet_mesh",
-    "load_obj_triangle_mesh",
-    "load_point_set",
-    "load_seg_line_mesh",
-    "set_accelerated_deformable_solve",
-]
-
-
-from collections.abc import Sequence
-import enum
-import os
-import pathlib
-from typing import Annotated, overload
-
-import numpy
-from numpy.typing import NDArray
-
-import dartpy.dynamics
-import dartpy.simulation
-import dartpy.simulation_experimental.diff as diff
-
-
-class JointType(enum.Enum):
-    FIXED = 0
-
-    REVOLUTE = 1
-
-    PRISMATIC = 2
-
-    SCREW = 3
-
-    UNIVERSAL = 4
-
-    SPHERICAL = 5
-
-    PLANAR = 6
-
-    FLOATING = 7
-
-    CUSTOM = 8
-
-class ActuatorType(enum.Enum):
-    FORCE = 0
-
-    PASSIVE = 1
-
-    SERVO = 2
-
-    VELOCITY = 3
-
-    ACCELERATION = 4
-
-    LOCKED = 5
-
-    MIMIC = 6
-
-class LoopClosureFamily(enum.Enum):
-    RIGID = 0
-
-    POINT = 1
-
-    DISTANCE = 2
-
-class LoopClosureResidualCoordinates(enum.Enum):
-    WORLD = 0
-
-class WorldSyncStage(enum.Enum):
-    KINEMATICS = 0
-
-class RigidBodySolver(enum.Enum):
-    SEQUENTIAL_IMPULSE = 0
-
-    IPC = 1
-
-class ContactSolverMethod(enum.Enum):
-    SEQUENTIAL_IMPULSE = 0
-
-    BOXED_LCP = 1
-
-class ContactGradientMode(enum.Enum):
-    ANALYTIC = 0
-
-    COMPLEMENTARITY_AWARE = 1
-
-    PRE_CONTACT_SURROGATE = 2
-
-class PhysicalParameter(enum.Enum):
-    MASS = 0
-
-    CENTER_OF_MASS = 1
-
-    INERTIA = 2
-
-    FRICTION = 3
-
-class CollisionShapeType(enum.Enum):
-    SPHERE = 0
-
-    BOX = 1
-
-    MESH = 2
-
-    CAPSULE = 3
-
-    CYLINDER = 4
-
-    PLANE = 5
-
-class CollisionShape:
-    @staticmethod
-    def sphere(radius: float, local_transform: object | None = ...) -> CollisionShape: ...
-
-    @staticmethod
-    def box(half_extents: object, local_transform: object | None = ...) -> CollisionShape: ...
-
-    @staticmethod
-    def capsule(radius: float, half_height: float, local_transform: object | None = ...) -> CollisionShape: ...
-
-    @staticmethod
-    def cylinder(radius: float, half_height: float, local_transform: object | None = ...) -> CollisionShape: ...
-
-    @staticmethod
-    def plane(normal: object, offset: float, local_transform: object | None = ...) -> CollisionShape: ...
-
-    @staticmethod
-    def mesh(vertices: object, triangles: object, local_transform: object | None = ...) -> CollisionShape: ...
-
-    @property
-    def type(self) -> CollisionShapeType: ...
-
-    @property
-    def radius(self) -> float: ...
-
-    @property
-    def height(self) -> float: ...
-
-    @property
-    def half_height(self) -> float: ...
-
-    @property
-    def half_extents(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def normal(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def offset(self) -> float: ...
-
-    @property
-    def vertices(self) -> list[Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]]: ...
-
-    @property
-    def triangles(self) -> list[Annotated[NDArray[numpy.int32], dict(shape=(3), order='C')]]: ...
-
-    @property
-    def local_transform(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-class ClosureKinematicsPolicy(enum.Enum):
-    RESIDUAL_ONLY = 0
-
-    PROJECT = 1
-
-class ClosureDynamicsPolicy(enum.Enum):
-    RESIDUAL_ONLY = 0
-
-    SOLVE = 1
-
-class StateVariable:
-    @property
-    def name(self) -> str: ...
-
-    @property
-    def start_index(self) -> int: ...
-
-    @property
-    def dimension(self) -> int: ...
-
-    @property
-    def lower_bound(self) -> float: ...
-
-    @property
-    def upper_bound(self) -> float: ...
-
-    def __repr__(self) -> str: ...
-
-class StateSpace:
-    def __init__(self) -> None: ...
-
-    def add_variable(self, name: str, dimension: int, *, lower: float = ..., upper: float = ...) -> StateSpace: ...
-
-    def add_variables(self, names: Sequence[str], *, lower: float = ..., upper: float = ...) -> StateSpace: ...
-
-    def finalize(self) -> None: ...
-
-    def has_variable(self, name: str) -> bool: ...
-
-    def get_variable(self, name: str) -> object: ...
-
-    def get_variable_index(self, name: str) -> object: ...
-
-    @property
-    def dimension(self) -> int: ...
-
-    @property
-    def num_variables(self) -> int: ...
-
-    @property
-    def is_finalized(self) -> bool: ...
-
-    @property
-    def variables(self) -> list[StateVariable]: ...
-
-    @property
-    def variable_names(self) -> list[str]: ...
-
-    @property
-    def lower_bounds(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @property
-    def upper_bounds(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    def __repr__(self) -> str: ...
-
-class JointSpec:
-    def __init__(self, name: str = ..., type: JointType = JointType.REVOLUTE, axis: object | None = ..., axis2: object | None = ..., transform_from_parent: object | None = ..., transform_to_parent: object | None = ...) -> None: ...
-
-    @property
-    def name(self) -> str: ...
-
-    @name.setter
-    def name(self, arg: str, /) -> None: ...
-
-    @property
-    def type(self) -> JointType: ...
-
-    @type.setter
-    def type(self, arg: JointType, /) -> None: ...
-
-    @property
-    def axis(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @axis.setter
-    def axis(self, arg: object, /) -> None: ...
-
-    @property
-    def axis2(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @axis2.setter
-    def axis2(self, arg: object, /) -> None: ...
-
-    @property
-    def transform_from_parent(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @transform_from_parent.setter
-    def transform_from_parent(self, arg: object, /) -> None: ...
-
-    @property
-    def transform_to_parent(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @transform_to_parent.setter
-    def transform_to_parent(self, arg: object, /) -> None: ...
-
-    def __repr__(self) -> str: ...
-
-class Frame:
-    @staticmethod
-    def world() -> Frame: ...
-
-    @overload
-    def relative_transform(self, relative_to: object) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @overload
-    def relative_transform(self, relative_to: object, expressed_in: object) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    def is_same_instance_as(self, other: object) -> bool: ...
-
-    @property
-    def name(self) -> str: ...
-
-    @property
-    def parent_frame(self) -> Frame: ...
-
-    @parent_frame.setter
-    def parent_frame(self, arg: object, /) -> None: ...
-
-    @property
-    def local_transform(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @property
-    def translation(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def rotation(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3, 3), order='F')]: ...
-
-    @property
-    def quaternion(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4), order='C')]: ...
-
-    @property
-    def transform(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @property
-    def is_valid(self) -> bool: ...
-
-    @property
-    def is_world(self) -> bool: ...
-
-    def __eq__(self, arg: object, /) -> bool: ...
-
-    def __ne__(self, arg: object, /) -> bool: ...
-
-    def __repr__(self) -> str: ...
-
-class FreeFrame(Frame):
-    @property
-    def local_transform(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @local_transform.setter
-    def local_transform(self, arg: object, /) -> None: ...
-
-    def __repr__(self) -> str: ...
-
-class FixedFrame(Frame):
-    @property
-    def local_transform(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @local_transform.setter
-    def local_transform(self, arg: object, /) -> None: ...
-
-    def __repr__(self) -> str: ...
-
-class Joint:
-    @property
-    def name(self) -> str: ...
-
-    @property
-    def type(self) -> JointType: ...
-
-    @property
-    def actuator_type(self) -> ActuatorType: ...
-
-    @actuator_type.setter
-    def actuator_type(self, arg: ActuatorType, /) -> None: ...
-
-    @property
-    def command_velocity(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @command_velocity.setter
-    def command_velocity(self, arg: object, /) -> None: ...
-
-    @property
-    def axis(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def axis2(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def pitch(self) -> float: ...
-
-    @pitch.setter
-    def pitch(self, arg: float, /) -> None: ...
-
-    @property
-    def num_dofs(self) -> int: ...
-
-    @property
-    def position(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @position.setter
-    def position(self, arg: object, /) -> None: ...
-
-    @property
-    def velocity(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @velocity.setter
-    def velocity(self, arg: object, /) -> None: ...
-
-    @property
-    def force(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @force.setter
-    def force(self, arg: object, /) -> None: ...
-
-    @property
-    def acceleration(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @property
-    def spring_stiffness(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @spring_stiffness.setter
-    def spring_stiffness(self, arg: object, /) -> None: ...
-
-    @property
-    def rest_position(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @rest_position.setter
-    def rest_position(self, arg: object, /) -> None: ...
-
-    @property
-    def damping_coefficient(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @damping_coefficient.setter
-    def damping_coefficient(self, arg: object, /) -> None: ...
-
-    @property
-    def armature(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @armature.setter
-    def armature(self, arg: object, /) -> None: ...
-
-    @property
-    def coulomb_friction(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @coulomb_friction.setter
-    def coulomb_friction(self, arg: object, /) -> None: ...
-
-    def set_position_limits(self, lower: object, upper: object) -> None: ...
-
-    @property
-    def position_lower_limits(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @property
-    def position_upper_limits(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    def set_velocity_limits(self, lower: object, upper: object) -> None: ...
-
-    @property
-    def velocity_lower_limits(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @property
-    def velocity_upper_limits(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    def set_effort_limits(self, lower: object, upper: object) -> None: ...
-
-    @property
-    def effort_lower_limits(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @property
-    def effort_upper_limits(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @property
-    def parent_link(self) -> Link: ...
-
-    @property
-    def child_link(self) -> Link: ...
-
-    @property
-    def parent_rigid_body(self) -> RigidBody: ...
-
-    @property
-    def child_rigid_body(self) -> RigidBody: ...
-
-    @property
-    def is_valid(self) -> bool: ...
-
-    def __repr__(self) -> str: ...
-
-class Link(Frame):
-    @property
-    def name(self) -> str: ...
-
-    @property
-    def parent_joint(self) -> Joint: ...
-
-    @property
-    def mass(self) -> float: ...
-
-    @mass.setter
-    def mass(self, arg: float, /) -> None: ...
-
-    @property
-    def inertia(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3, 3), order='F')]: ...
-
-    @inertia.setter
-    def inertia(self, arg: object, /) -> None: ...
-
-    @property
-    def center_of_mass(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @center_of_mass.setter
-    def center_of_mass(self, arg: object, /) -> None: ...
-
-    def apply_force(self, force: object, point: object = ..., force_in_world_frame: bool = ..., point_in_world_frame: bool = ...) -> None: ...
-
-    @property
-    def translation(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def rotation(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3, 3), order='F')]: ...
-
-    @property
-    def quaternion(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4), order='C')]: ...
-
-    @property
-    def transform(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    def set_collision_shape(self, shape: CollisionShape) -> None: ...
-
-    def add_collision_shape(self, shape: CollisionShape) -> None: ...
-
-    @property
-    def collision_shape(self) -> CollisionShape | None: ...
-
-    @property
-    def collision_shapes(self) -> list[CollisionShape]: ...
-
-    @property
-    def has_collision_shape(self) -> bool: ...
-
-    @property
-    def is_valid(self) -> bool: ...
-
-    def __repr__(self) -> str: ...
-
-class LoopClosure:
-    def compute_residual(self) -> LoopClosureResidual: ...
-
-    @property
-    def name(self) -> str: ...
-
-    @property
-    def family(self) -> LoopClosureFamily: ...
-
-    @property
-    def frame_a(self) -> Frame: ...
-
-    @property
-    def frame_b(self) -> Frame: ...
-
-    @property
-    def offset_a(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @property
-    def offset_b(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @property
-    def runtime_policy(self) -> LoopClosureRuntimePolicy: ...
-
-    @runtime_policy.setter
-    def runtime_policy(self, arg: LoopClosureRuntimePolicy, /) -> None: ...
-
-    @property
-    def enabled(self) -> bool: ...
-
-    @enabled.setter
-    def enabled(self, arg: bool, /) -> None: ...
-
-    @property
-    def kinematics(self) -> ClosureKinematicsPolicy: ...
-
-    @kinematics.setter
-    def kinematics(self, arg: ClosureKinematicsPolicy, /) -> None: ...
-
-    @property
-    def dynamics(self) -> ClosureDynamicsPolicy: ...
-
-    @dynamics.setter
-    def dynamics(self, arg: ClosureDynamicsPolicy, /) -> None: ...
-
-    @property
-    def is_valid(self) -> bool: ...
-
-    def __repr__(self) -> str: ...
-
-class RigidBody(Frame):
-    def apply_force(self, force: object) -> None: ...
-
-    def clear_force(self) -> None: ...
-
-    def apply_torque(self, torque: object) -> None: ...
-
-    def clear_torque(self) -> None: ...
-
-    @property
-    def name(self) -> str: ...
-
-    @property
-    def translation(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def rotation(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3, 3), order='F')]: ...
-
-    @property
-    def quaternion(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4), order='C')]: ...
-
-    @property
-    def transform(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @transform.setter
-    def transform(self, arg: object, /) -> None: ...
-
-    @property
-    def linear_velocity(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @linear_velocity.setter
-    def linear_velocity(self, arg: object, /) -> None: ...
-
-    @property
-    def angular_velocity(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @angular_velocity.setter
-    def angular_velocity(self, arg: object, /) -> None: ...
-
-    @property
-    def mass(self) -> float: ...
-
-    @mass.setter
-    def mass(self, arg: float, /) -> None: ...
-
-    @property
-    def inertia(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3, 3), order='F')]: ...
-
-    @inertia.setter
-    def inertia(self, arg: object, /) -> None: ...
-
-    @property
-    def force(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @force.setter
-    def force(self, arg: object, /) -> None: ...
-
-    @property
-    def torque(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @torque.setter
-    def torque(self, arg: object, /) -> None: ...
-
-    @property
-    def is_static(self) -> bool: ...
-
-    @is_static.setter
-    def is_static(self, arg: bool, /) -> None: ...
-
-    @property
-    def is_kinematic(self) -> bool: ...
-
-    @is_kinematic.setter
-    def is_kinematic(self, arg: bool, /) -> None: ...
-
-    @property
-    def restitution(self) -> float: ...
-
-    @restitution.setter
-    def restitution(self, arg: float, /) -> None: ...
-
-    @property
-    def friction(self) -> float: ...
-
-    @friction.setter
-    def friction(self, arg: float, /) -> None: ...
-
-    def set_collision_shape(self, shape: CollisionShape) -> None: ...
-
-    def add_collision_shape(self, shape: CollisionShape) -> None: ...
-
-    @property
-    def is_deformable_surface_ccd_obstacle(self) -> bool: ...
-
-    @is_deformable_surface_ccd_obstacle.setter
-    def is_deformable_surface_ccd_obstacle(self, arg: bool, /) -> None: ...
-
-    @property
-    def is_deformable_obstacle_barrier_only(self) -> bool: ...
-
-    @is_deformable_obstacle_barrier_only.setter
-    def is_deformable_obstacle_barrier_only(self, arg: bool, /) -> None: ...
-
-    @property
-    def is_deformable_ground_barrier(self) -> bool: ...
-
-    @is_deformable_ground_barrier.setter
-    def is_deformable_ground_barrier(self, arg: bool, /) -> None: ...
-
-    @property
-    def collision_shape(self) -> CollisionShape | None: ...
-
-    @property
-    def collision_shapes(self) -> list[CollisionShape]: ...
-
-    @property
-    def has_collision_shape(self) -> bool: ...
-
-    @property
-    def linear_momentum(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def angular_momentum(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def kinetic_energy(self) -> float: ...
-
-    @property
-    def potential_energy(self) -> float: ...
-
-    def __repr__(self) -> str: ...
-
-class LoopClosureSpec:
-    def __init__(self, frame_a: object, frame_b: object, family: LoopClosureFamily = LoopClosureFamily.RIGID, *, offset_a: object | None = ..., offset_b: object | None = ..., distance: float = ...) -> None: ...
-
-    @property
-    def frame_a(self) -> Frame: ...
-
-    @frame_a.setter
-    def frame_a(self, arg: object, /) -> None: ...
-
-    @property
-    def frame_b(self) -> Frame: ...
-
-    @frame_b.setter
-    def frame_b(self, arg: object, /) -> None: ...
-
-    @property
-    def family(self) -> LoopClosureFamily: ...
-
-    @family.setter
-    def family(self, arg: LoopClosureFamily, /) -> None: ...
-
-    @property
-    def distance(self) -> float: ...
-
-    @distance.setter
-    def distance(self, arg: float, /) -> None: ...
-
-    @property
-    def offset_a(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @offset_a.setter
-    def offset_a(self, arg: object, /) -> None: ...
-
-    @property
-    def offset_b(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4, 4), order='F')]: ...
-
-    @offset_b.setter
-    def offset_b(self, arg: object, /) -> None: ...
-
-    def __repr__(self) -> str: ...
-
-class LoopClosureRuntimePolicy:
-    def __init__(self, enabled: bool = ..., kinematics: ClosureKinematicsPolicy = ClosureKinematicsPolicy.RESIDUAL_ONLY, dynamics: ClosureDynamicsPolicy = ClosureDynamicsPolicy.RESIDUAL_ONLY) -> None: ...
-
-    @property
-    def enabled(self) -> bool: ...
-
-    @enabled.setter
-    def enabled(self, arg: bool, /) -> None: ...
-
-    @property
-    def kinematics(self) -> ClosureKinematicsPolicy: ...
-
-    @kinematics.setter
-    def kinematics(self, arg: ClosureKinematicsPolicy, /) -> None: ...
-
-    @property
-    def dynamics(self) -> ClosureDynamicsPolicy: ...
-
-    @dynamics.setter
-    def dynamics(self, arg: ClosureDynamicsPolicy, /) -> None: ...
-
-    def __repr__(self) -> str: ...
-
-class MultibodyOptions:
-    def __init__(self, integration_family: str = ...) -> None: ...
-
-    @property
-    def integration_family(self) -> str: ...
-
-    @integration_family.setter
-    def integration_family(self, arg: str, /) -> None: ...
-
-    def __repr__(self) -> str: ...
-
-class LoopClosureResidual:
-    @property
-    def value(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @property
-    def norm(self) -> float: ...
-
-    @property
-    def enabled(self) -> bool: ...
-
-    @property
-    def active(self) -> bool: ...
-
-    @property
-    def coordinates(self) -> LoopClosureResidualCoordinates: ...
-
-    @property
-    def force_available(self) -> bool: ...
-
-    def __repr__(self) -> str: ...
-
-class Multibody:
-    @overload
-    def add_link(self, name: str = ...) -> Link: ...
-
-    @overload
-    def add_link(self, name: str, *, parent: Link, joint: JointSpec = ...) -> Link: ...
-
-    def get_link(self, name: str) -> object: ...
-
-    def get_joint(self, name: str) -> object: ...
-
-    def set_ground_contact(self, plane_normal: Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')], plane_point: Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')], stiffness: float, friction_coefficient: float = ..., friction_regularization: float = ..., damping_coefficient: float = ..., dual_update_cadence: int = ...) -> None: ...
-
-    def add_ground_contact_point(self, link: Link, local_point: Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]) -> None: ...
-
-    @property
-    def name(self) -> str: ...
-
-    @name.setter
-    def name(self, arg: str, /) -> None: ...
-
-    @property
-    def num_links(self) -> int: ...
-
-    @property
-    def num_joints(self) -> int: ...
-
-    @property
-    def num_dofs(self) -> int: ...
-
-    @property
-    def links(self) -> list[Link]: ...
-
-    @property
-    def joints(self) -> list[Joint]: ...
-
-    @property
-    def link_names(self) -> list[str]: ...
-
-    @property
-    def joint_names(self) -> list[str]: ...
-
-    @property
-    def mass_matrix(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None, None), order='F')]: ...
-
-    @property
-    def inverse_mass_matrix(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None, None), order='F')]: ...
-
-    @property
-    def coriolis_forces(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @property
-    def gravity_forces(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @property
-    def coriolis_and_gravity_forces(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    def compute_inverse_dynamics(self, desired_acceleration: object) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    def compute_impulse_response(self, joint_impulse: object) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    def get_jacobian(self, link: Link) -> Annotated[NDArray[numpy.float64], dict(shape=(None, None), order='F')]: ...
-
-    def get_world_jacobian(self, link: Link) -> Annotated[NDArray[numpy.float64], dict(shape=(None, None), order='F')]: ...
-
-    @property
-    def is_valid(self) -> bool: ...
-
-    def __repr__(self) -> str: ...
-
-class RigidBodyOptions:
-    def __init__(self, mass: float = ..., position: object | None = ..., orientation: object | None = ..., linear_velocity: object | None = ..., angular_velocity: object | None = ..., inertia: object | None = ...) -> None: ...
-
-    @property
-    def mass(self) -> float: ...
-
-    @mass.setter
-    def mass(self, arg: float, /) -> None: ...
-
-    @property
-    def position(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @position.setter
-    def position(self, arg: object, /) -> None: ...
-
-    @property
-    def orientation(self) -> Annotated[NDArray[numpy.float64], dict(shape=(4), order='C')]: ...
-
-    @orientation.setter
-    def orientation(self, arg: object, /) -> None: ...
-
-    @property
-    def linear_velocity(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @linear_velocity.setter
-    def linear_velocity(self, arg: object, /) -> None: ...
-
-    @property
-    def angular_velocity(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @angular_velocity.setter
-    def angular_velocity(self, arg: object, /) -> None: ...
-
-    @property
-    def inertia(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3, 3), order='F')]: ...
-
-    @inertia.setter
-    def inertia(self, arg: object, /) -> None: ...
-
-    @property
-    def is_static(self) -> bool: ...
-
-    @is_static.setter
-    def is_static(self, arg: bool, /) -> None: ...
-
-    def __repr__(self) -> str: ...
-
-class CollisionBody:
-    @property
-    def name(self) -> str: ...
-
-    @property
-    def is_rigid_body(self) -> bool: ...
-
-    @property
-    def is_link(self) -> bool: ...
-
-    @property
-    def is_valid(self) -> bool: ...
-
-    def as_rigid_body(self) -> object: ...
-
-    def as_link(self) -> object: ...
-
-class Contact:
-    @property
-    def body_a(self) -> CollisionBody: ...
-
-    @property
-    def body_b(self) -> CollisionBody: ...
-
-    @property
-    def point(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def normal(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def depth(self) -> float: ...
-
-    @property
-    def shape_index_a(self) -> int: ...
-
-    @property
-    def shape_index_b(self) -> int: ...
-
-    @property
-    def local_point_a(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @property
-    def local_point_b(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-class CollisionQueryOptions:
-    def __init__(self, include_same_multibody_link_pairs: bool = ..., *, include_rigid_body_pairs: bool = ..., include_rigid_body_link_pairs: bool = ..., include_link_pairs: bool = ...) -> None: ...
-
-    @property
-    def include_same_multibody_link_pairs(self) -> bool: ...
-
-    @include_same_multibody_link_pairs.setter
-    def include_same_multibody_link_pairs(self, arg: bool, /) -> None: ...
-
-    @property
-    def include_rigid_body_pairs(self) -> bool: ...
-
-    @include_rigid_body_pairs.setter
-    def include_rigid_body_pairs(self, arg: bool, /) -> None: ...
-
-    @property
-    def include_rigid_body_link_pairs(self) -> bool: ...
-
-    @include_rigid_body_link_pairs.setter
-    def include_rigid_body_link_pairs(self, arg: bool, /) -> None: ...
-
-    @property
-    def include_link_pairs(self) -> bool: ...
-
-    @include_link_pairs.setter
-    def include_link_pairs(self, arg: bool, /) -> None: ...
-
-    def __repr__(self) -> str: ...
-
-class StepDerivatives:
-    @property
-    def state_jacobian(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None, None), order='F')]: ...
-
-    @property
-    def control_jacobian(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None, None), order='F')]: ...
-
-    @property
-    def parameter_jacobian(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None, None), order='F')]: ...
-
-    def __repr__(self) -> str: ...
-
-class StepGradient:
-    @property
-    def state(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @property
-    def control(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    def __repr__(self) -> str: ...
-
-class DeformableSolverDiagnostics:
-    @property
-    def body_count(self) -> int: ...
-
-    @property
-    def node_count(self) -> int: ...
-
-    @property
-    def edge_count(self) -> int: ...
-
-    @property
-    def solver_iterations(self) -> int: ...
-
-    @property
-    def objective_evaluations(self) -> int: ...
-
-    @property
-    def line_search_trials(self) -> int: ...
-
-    @property
-    def projected_newton_steps(self) -> int: ...
-
-    @property
-    def projected_newton_fallbacks(self) -> int: ...
-
-    @property
-    def projected_newton_hessian_nonzeros(self) -> int: ...
-
-    @property
-    def projected_newton_hessian_storage_bytes(self) -> int: ...
-
-    @property
-    def projected_newton_iterative_solves(self) -> int: ...
-
-    @property
-    def projected_newton_matrix_free_solves(self) -> int: ...
-
-    @property
-    def projected_newton_iterative_iterations(self) -> int: ...
-
-    @property
-    def projected_newton_iterative_max_error(self) -> float: ...
-
-    @property
-    def self_contact_barrier_active_contacts(self) -> int: ...
-
-    @property
-    def friction_dissipation(self) -> float: ...
-
-    @property
-    def min_active_contact_distance(self) -> float: ...
-
-    @property
-    def converged_active_contact_count(self) -> int: ...
-
-class DeformableMaterialProperties:
-    def __init__(self) -> None: ...
-
-    @property
-    def density(self) -> float: ...
-
-    @density.setter
-    def density(self, arg: float, /) -> None: ...
-
-    @property
-    def youngs_modulus(self) -> float: ...
-
-    @youngs_modulus.setter
-    def youngs_modulus(self, arg: float, /) -> None: ...
-
-    @property
-    def poisson_ratio(self) -> float: ...
-
-    @poisson_ratio.setter
-    def poisson_ratio(self, arg: float, /) -> None: ...
-
-    @property
-    def friction_coefficient(self) -> float: ...
-
-    @friction_coefficient.setter
-    def friction_coefficient(self, arg: float, /) -> None: ...
-
-    @property
-    def use_finite_element_elasticity(self) -> bool: ...
-
-    @use_finite_element_elasticity.setter
-    def use_finite_element_elasticity(self, arg: bool, /) -> None: ...
-
-    @property
-    def use_fixed_corotational_elasticity(self) -> bool: ...
-
-    @use_fixed_corotational_elasticity.setter
-    def use_fixed_corotational_elasticity(self, arg: bool, /) -> None: ...
-
-    @property
-    def use_adaptive_barrier_stiffness(self) -> bool: ...
-
-    @use_adaptive_barrier_stiffness.setter
-    def use_adaptive_barrier_stiffness(self, arg: bool, /) -> None: ...
-
-    @property
-    def use_iterative_linear_solver(self) -> bool: ...
-
-    @use_iterative_linear_solver.setter
-    def use_iterative_linear_solver(self, arg: bool, /) -> None: ...
-
-    @property
-    def use_matrix_free_linear_solver(self) -> bool: ...
-
-    @use_matrix_free_linear_solver.setter
-    def use_matrix_free_linear_solver(self, arg: bool, /) -> None: ...
-
-class DeformableEdge:
-    def __init__(self, node_a: int = ..., node_b: int = ..., rest_length: float = ...) -> None: ...
-
-    @property
-    def node_a(self) -> int: ...
-
-    @node_a.setter
-    def node_a(self, arg: int, /) -> None: ...
-
-    @property
-    def node_b(self) -> int: ...
-
-    @node_b.setter
-    def node_b(self, arg: int, /) -> None: ...
-
-    @property
-    def rest_length(self) -> float: ...
-
-    @rest_length.setter
-    def rest_length(self, arg: float, /) -> None: ...
-
-class DeformableSurfaceTriangle:
-    def __init__(self, node_a: int = ..., node_b: int = ..., node_c: int = ...) -> None: ...
-
-    @property
-    def node_a(self) -> int: ...
-
-    @node_a.setter
-    def node_a(self, arg: int, /) -> None: ...
-
-    @property
-    def node_b(self) -> int: ...
-
-    @node_b.setter
-    def node_b(self, arg: int, /) -> None: ...
-
-    @property
-    def node_c(self) -> int: ...
-
-    @node_c.setter
-    def node_c(self, arg: int, /) -> None: ...
-
-class DeformableTetrahedron:
-    def __init__(self, node_a: int = ..., node_b: int = ..., node_c: int = ..., node_d: int = ...) -> None: ...
-
-    @property
-    def node_a(self) -> int: ...
-
-    @node_a.setter
-    def node_a(self, arg: int, /) -> None: ...
-
-    @property
-    def node_b(self) -> int: ...
-
-    @node_b.setter
-    def node_b(self, arg: int, /) -> None: ...
-
-    @property
-    def node_c(self) -> int: ...
-
-    @node_c.setter
-    def node_c(self, arg: int, /) -> None: ...
-
-    @property
-    def node_d(self) -> int: ...
-
-    @node_d.setter
-    def node_d(self, arg: int, /) -> None: ...
-
-class DeformableDirichletBoundaryCondition:
-    def __init__(self) -> None: ...
-
-    @property
-    def nodes(self) -> list[int]: ...
-
-    @nodes.setter
-    def nodes(self, arg: Sequence[int], /) -> None: ...
-
-    @property
-    def linear_velocity(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @linear_velocity.setter
-    def linear_velocity(self, arg: Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')], /) -> None: ...
-
-    @property
-    def angular_velocity(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @angular_velocity.setter
-    def angular_velocity(self, arg: Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')], /) -> None: ...
-
-    @property
-    def center(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @center.setter
-    def center(self, arg: Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')], /) -> None: ...
-
-    @property
-    def start_time(self) -> float: ...
-
-    @start_time.setter
-    def start_time(self, arg: float, /) -> None: ...
-
-    @property
-    def end_time(self) -> float: ...
-
-    @end_time.setter
-    def end_time(self, arg: float, /) -> None: ...
-
-class DeformableNeumannBoundaryCondition:
-    def __init__(self) -> None: ...
-
-    @property
-    def nodes(self) -> list[int]: ...
-
-    @nodes.setter
-    def nodes(self, arg: Sequence[int], /) -> None: ...
-
-    @property
-    def acceleration(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @acceleration.setter
-    def acceleration(self, arg: Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')], /) -> None: ...
-
-    @property
-    def start_time(self) -> float: ...
-
-    @start_time.setter
-    def start_time(self, arg: float, /) -> None: ...
-
-    @property
-    def end_time(self) -> float: ...
-
-    @end_time.setter
-    def end_time(self, arg: float, /) -> None: ...
-
-class DeformableBodyOptions:
-    def __init__(self) -> None: ...
-
-    @property
-    def positions(self) -> list[Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]]: ...
-
-    @positions.setter
-    def positions(self, arg: Sequence[Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]], /) -> None: ...
-
-    @property
-    def velocities(self) -> list[Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]]: ...
-
-    @velocities.setter
-    def velocities(self, arg: Sequence[Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]], /) -> None: ...
-
-    @property
-    def masses(self) -> list[float]: ...
-
-    @masses.setter
-    def masses(self, arg: Sequence[float], /) -> None: ...
-
-    @property
-    def edges(self) -> list[DeformableEdge]: ...
-
-    @edges.setter
-    def edges(self, arg: Sequence[DeformableEdge], /) -> None: ...
-
-    @property
-    def surface_triangles(self) -> list[DeformableSurfaceTriangle]: ...
-
-    @surface_triangles.setter
-    def surface_triangles(self, arg: Sequence[DeformableSurfaceTriangle], /) -> None: ...
-
-    @property
-    def tetrahedra(self) -> list[DeformableTetrahedron]: ...
-
-    @tetrahedra.setter
-    def tetrahedra(self, arg: Sequence[DeformableTetrahedron], /) -> None: ...
-
-    @property
-    def dirichlet_boundary_conditions(self) -> list[DeformableDirichletBoundaryCondition]: ...
-
-    @dirichlet_boundary_conditions.setter
-    def dirichlet_boundary_conditions(self, arg: Sequence[DeformableDirichletBoundaryCondition], /) -> None: ...
-
-    @property
-    def neumann_boundary_conditions(self) -> list[DeformableNeumannBoundaryCondition]: ...
-
-    @neumann_boundary_conditions.setter
-    def neumann_boundary_conditions(self, arg: Sequence[DeformableNeumannBoundaryCondition], /) -> None: ...
-
-    @property
-    def fixed_nodes(self) -> list[int]: ...
-
-    @fixed_nodes.setter
-    def fixed_nodes(self, arg: Sequence[int], /) -> None: ...
-
-    @property
-    def edge_stiffness(self) -> float: ...
-
-    @edge_stiffness.setter
-    def edge_stiffness(self, arg: float, /) -> None: ...
-
-    @property
-    def damping(self) -> float: ...
-
-    @damping.setter
-    def damping(self, arg: float, /) -> None: ...
-
-    @property
-    def material(self) -> DeformableMaterialProperties: ...
-
-    @material.setter
-    def material(self, arg: DeformableMaterialProperties, /) -> None: ...
-
-class DeformableSolverOptions:
-    def __init__(self) -> None: ...
-
-    @property
-    def iterations(self) -> int: ...
-
-    @iterations.setter
-    def iterations(self, arg: int, /) -> None: ...
-
-    @property
-    def convergence_tolerance(self) -> float: ...
-
-    @convergence_tolerance.setter
-    def convergence_tolerance(self, arg: float, /) -> None: ...
-
-    @property
-    def use_acceleration(self) -> bool: ...
-
-    @use_acceleration.setter
-    def use_acceleration(self, arg: bool, /) -> None: ...
-
-    @property
-    def acceleration_spectral_radius(self) -> float: ...
-
-    @acceleration_spectral_radius.setter
-    def acceleration_spectral_radius(self, arg: float, /) -> None: ...
-
-    @property
-    def stiffness_damping(self) -> float: ...
-
-    @stiffness_damping.setter
-    def stiffness_damping(self, arg: float, /) -> None: ...
-
-    @property
-    def worker_threads(self) -> int: ...
-
-    @worker_threads.setter
-    def worker_threads(self, arg: int, /) -> None: ...
-
-    @property
-    def ground_contact_stiffness(self) -> float: ...
-
-    @ground_contact_stiffness.setter
-    def ground_contact_stiffness(self, arg: float, /) -> None: ...
-
-class DeformableBody:
-    @property
-    def is_valid(self) -> bool: ...
-
-    @property
-    def name(self) -> str: ...
-
-    @property
-    def node_count(self) -> int: ...
-
-    @property
-    def edge_count(self) -> int: ...
-
-    @property
-    def surface_triangle_count(self) -> int: ...
-
-    @property
-    def tetrahedron_count(self) -> int: ...
-
-    def node_position(self, node: int) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    def set_node_position(self, node: int, position: Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]) -> None: ...
-
-    def node_velocity(self, node: int) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    def set_node_velocity(self, node: int, velocity: Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]) -> None: ...
-
-    def node_mass(self, node: int) -> float: ...
-
-    def is_fixed_node(self, node: int) -> bool: ...
-
-    def edge(self, edge: int) -> DeformableEdge: ...
-
-    def surface_triangle(self, triangle: int) -> DeformableSurfaceTriangle: ...
-
-    def tetrahedron(self, tetrahedron: int) -> DeformableTetrahedron: ...
-
-    def tetrahedron_rest_volume(self, tetrahedron: int) -> float: ...
-
-    @property
-    def material_properties(self) -> DeformableMaterialProperties: ...
-
-class SkeletonLoadOptions:
-    def __init__(self) -> None: ...
-
-    @property
-    def root_anchor_prefix(self) -> str: ...
-
-    @root_anchor_prefix.setter
-    def root_anchor_prefix(self, arg: str, /) -> None: ...
-
-class ModelFormat(enum.Enum):
-    AUTO = 0
-
-    SKEL = 1
-
-    SDF = 2
-
-    URDF = 3
-
-    MJCF = 4
-
-class RootJointType(enum.Enum):
-    FLOATING = 0
-
-    FIXED = 1
-
-class ReadOptions:
-    def __init__(self) -> None: ...
-
-    @property
-    def format(self) -> ModelFormat: ...
-
-    @format.setter
-    def format(self, arg: ModelFormat, /) -> None: ...
-
-    @property
-    def sdf_default_root_joint_type(self) -> RootJointType: ...
-
-    @sdf_default_root_joint_type.setter
-    def sdf_default_root_joint_type(self, arg: RootJointType, /) -> None: ...
-
-    def add_package_directory(self, package_name: str, package_directory: str) -> None: ...
-
-class DeformableSceneLoadOptions:
-    def __init__(self) -> None: ...
-
-    @property
-    def asset_root(self) -> pathlib.Path: ...
-
-    @asset_root.setter
-    def asset_root(self, arg: str | os.PathLike, /) -> None: ...
-
-    @property
-    def body_name_prefix(self) -> str: ...
-
-    @body_name_prefix.setter
-    def body_name_prefix(self, arg: str, /) -> None: ...
-
-    @property
-    def add_structural_springs(self) -> bool: ...
-
-    @add_structural_springs.setter
-    def add_structural_springs(self, arg: bool, /) -> None: ...
-
-    @property
-    def structural_spring_stiffness(self) -> float: ...
-
-    @structural_spring_stiffness.setter
-    def structural_spring_stiffness(self, arg: float, /) -> None: ...
-
-    @property
-    def damping(self) -> float: ...
-
-    @damping.setter
-    def damping(self, arg: float, /) -> None: ...
-
-    @property
-    def ignore_contact_directives(self) -> bool: ...
-
-    @ignore_contact_directives.setter
-    def ignore_contact_directives(self, arg: bool, /) -> None: ...
-
-class DeformableSceneBodyInfo:
-    @property
-    def name(self) -> str: ...
-
-    @property
-    def body(self) -> DeformableBody: ...
-
-    @property
-    def node_count(self) -> int: ...
-
-    @property
-    def tetrahedron_count(self) -> int: ...
-
-    @property
-    def surface_triangle_count(self) -> int: ...
-
-    @property
-    def dirichlet_condition_count(self) -> int: ...
-
-    @property
-    def neumann_condition_count(self) -> int: ...
-
-class DeformableSceneInfo:
-    @property
-    def duration(self) -> float: ...
-
-    @property
-    def time_step(self) -> float: ...
-
-    @property
-    def gravity_enabled(self) -> bool: ...
-
-    @property
-    def bodies(self) -> list[DeformableSceneBodyInfo]: ...
-
-    @property
-    def warnings(self) -> list[str]: ...
-
-class DeformableSceneDiagnostics:
-    @property
-    def frame(self) -> int: ...
-
-    @property
-    def time(self) -> float: ...
-
-    @property
-    def body_count(self) -> int: ...
-
-    @property
-    def node_count(self) -> int: ...
-
-    @property
-    def tetrahedron_count(self) -> int: ...
-
-    @property
-    def surface_triangle_count(self) -> int: ...
-
-    @property
-    def dirichlet_condition_count(self) -> int: ...
-
-    @property
-    def neumann_condition_count(self) -> int: ...
-
-    @property
-    def total_mass(self) -> float: ...
-
-    @property
-    def max_displacement(self) -> float: ...
-
-    @property
-    def min_z(self) -> float: ...
-
-    @property
-    def max_z(self) -> float: ...
-
-class SkeletonToMultibodyOptions:
-    def __init__(self) -> None: ...
-
-    @property
-    def name(self) -> str: ...
-
-    @name.setter
-    def name(self, arg: str, /) -> None: ...
-
-    @property
-    def base_link_name(self) -> str: ...
-
-    @base_link_name.setter
-    def base_link_name(self, arg: str, /) -> None: ...
-
-    @property
-    def copy_state(self) -> bool: ...
-
-    @copy_state.setter
-    def copy_state(self, arg: bool, /) -> None: ...
-
-    @property
-    def copy_joint_properties(self) -> bool: ...
-
-    @copy_joint_properties.setter
-    def copy_joint_properties(self, arg: bool, /) -> None: ...
-
-    @property
-    def load_collision_shapes(self) -> bool: ...
-
-    @load_collision_shapes.setter
-    def load_collision_shapes(self, arg: bool, /) -> None: ...
-
-def build_multibody_from_skeleton(world: World, skeleton: dartpy.dynamics.Skeleton, options: SkeletonToMultibodyOptions = ...) -> Multibody: ...
-
-def build_multibodies_from_world(world: World, legacy_world: dartpy.simulation.World, options: SkeletonToMultibodyOptions = ...) -> list[Multibody]: ...
-
-@overload
-def add_skeleton(world: World, skeleton: dartpy.dynamics.Skeleton, options: SkeletonLoadOptions = ...) -> Multibody: ...
-
-@overload
-def add_skeleton(world: World, uri: str, options: SkeletonLoadOptions = ...) -> Multibody: ...
-
-@overload
-def add_skeleton(world: World, uri: str, read_options: ReadOptions, options: SkeletonLoadOptions = ...) -> Multibody: ...
-
-@overload
-def add_world(world: object, source_world: dartpy.simulation.World, options: SkeletonLoadOptions = ...) -> list: ...
-
-@overload
-def add_world(world: object, uri: str, options: SkeletonLoadOptions = ...) -> list: ...
-
-@overload
-def add_world(world: object, uri: str, read_options: ReadOptions, options: SkeletonLoadOptions = ...) -> list: ...
-
-def load_deformable_scene(world: World, scene_path: str | os.PathLike, options: DeformableSceneLoadOptions = ...) -> DeformableSceneInfo: ...
-
-def collect_deformable_scene_diagnostics(world: World) -> DeformableSceneDiagnostics: ...
-
-def load_gmsh_tet_mesh(path: str | os.PathLike) -> DeformableBodyOptions: ...
-
-def load_obj_triangle_mesh(path: str | os.PathLike) -> DeformableBodyOptions: ...
-
-def load_seg_line_mesh(path: str | os.PathLike) -> DeformableBodyOptions: ...
-
-def load_point_set(path: str | os.PathLike) -> DeformableBodyOptions: ...
-
-class World:
-    def __init__(self, time_step: float = ..., *, gravity: object | None = ..., differentiable: bool = ..., contact_solver_method: ContactSolverMethod = ContactSolverMethod.SEQUENTIAL_IMPULSE, contact_gradient_mode: ContactGradientMode = ContactGradientMode.ANALYTIC) -> None: ...
-
-    def add_free_frame(self, name: str = ..., *, parent: object | None = ...) -> FreeFrame: ...
-
-    def add_fixed_frame(self, name: str, parent: object, offset: object | None = ...) -> FixedFrame: ...
-
-    def add_multibody(self, name: str) -> Multibody: ...
-
-    def get_multibody(self, name: str) -> object: ...
-
-    def has_multibody(self, name: str) -> bool: ...
-
-    def add_deformable_body(self, name: str, options: DeformableBodyOptions) -> DeformableBody: ...
-
-    def get_deformable_body(self, name: str) -> object: ...
-
-    def has_deformable_body(self, name: str) -> bool: ...
-
-    @property
-    def num_deformable_bodies(self) -> int: ...
-
-    def configure_deformable_solver(self, name: str, options: DeformableSolverOptions) -> None: ...
-
-    @overload
-    def add_loop_closure(self, spec: LoopClosureSpec, *, name: object | None = ...) -> LoopClosure: ...
-
-    @overload
-    def add_loop_closure(self, name: object, spec: LoopClosureSpec) -> LoopClosure: ...
-
-    @overload
-    def add_loop_closure(self, name: object | None = ..., *, frame_a: object, frame_b: object, family: LoopClosureFamily = LoopClosureFamily.RIGID, offset_a: object | None = ..., offset_b: object | None = ..., distance: float = ...) -> LoopClosure: ...
-
-    def has_loop_closure(self, name: str) -> bool: ...
-
-    def get_loop_closure(self, name: str) -> object: ...
-
-    def add_rigid_body(self, name: str, options: RigidBodyOptions = ..., *, mass: object | None = ..., position: object | None = ..., orientation: object | None = ..., linear_velocity: object | None = ..., angular_velocity: object | None = ..., inertia: object | None = ...) -> RigidBody: ...
-
-    def add_rigid_body_fixed_joint(self, name: str, parent: RigidBody, child: RigidBody) -> Joint: ...
-
-    def add_rigid_body_revolute_joint(self, name: str, parent: RigidBody, child: RigidBody, axis: object = ...) -> Joint: ...
-
-    def add_rigid_body_prismatic_joint(self, name: str, parent: RigidBody, child: RigidBody, axis: object = ...) -> Joint: ...
-
-    def get_rigid_body_joint(self, name: str) -> object: ...
-
-    def has_rigid_body_joint(self, name: str) -> bool: ...
-
-    def get_rigid_body_joints(self) -> list: ...
-
-    def get_rigid_body_fixed_joint(self, name: str) -> object: ...
-
-    def has_rigid_body_fixed_joint(self, name: str) -> bool: ...
-
-    def get_rigid_body_fixed_joints(self) -> list: ...
-
-    def has_rigid_body(self, name: str) -> bool: ...
-
-    def get_rigid_body(self, name: str) -> object: ...
-
-    @property
-    def is_simulation_mode(self) -> bool: ...
-
-    def enter_simulation_mode(self) -> None: ...
-
-    def update_kinematics(self) -> None: ...
-
-    def sync(self, stage: WorldSyncStage = WorldSyncStage.KINEMATICS) -> None: ...
-
-    def step(self, n: int = ...) -> None: ...
-
-    @property
-    def last_deformable_solver_diagnostics(self) -> DeformableSolverDiagnostics: ...
-
-    @property
-    def time_step(self) -> float: ...
-
-    @time_step.setter
-    def time_step(self, arg: float, /) -> None: ...
-
-    @property
-    def time(self) -> float: ...
-
-    @time.setter
-    def time(self, arg: float, /) -> None: ...
-
-    @property
-    def gravity(self) -> Annotated[NDArray[numpy.float64], dict(shape=(3), order='C')]: ...
-
-    @gravity.setter
-    def gravity(self, arg: object, /) -> None: ...
-
-    @property
-    def rigid_body_solver(self) -> RigidBodySolver: ...
-
-    @rigid_body_solver.setter
-    def rigid_body_solver(self, arg: RigidBodySolver, /) -> None: ...
-
-    @property
-    def multibody_options(self) -> MultibodyOptions: ...
-
-    @multibody_options.setter
-    def multibody_options(self, arg: MultibodyOptions, /) -> None: ...
-
-    @property
-    def frame(self) -> int: ...
-
-    @property
-    def num_multibodies(self) -> int: ...
-
-    @property
-    def num_loop_closures(self) -> int: ...
-
-    @property
-    def num_rigid_bodies(self) -> int: ...
-
-    @property
-    def num_rigid_body_joints(self) -> int: ...
-
-    @property
-    def num_rigid_body_fixed_joints(self) -> int: ...
-
-    @property
-    def is_differentiable(self) -> bool: ...
-
-    @property
-    def contact_solver_method(self) -> ContactSolverMethod: ...
-
-    @property
-    def contact_gradient_mode(self) -> ContactGradientMode: ...
-
-    @contact_gradient_mode.setter
-    def contact_gradient_mode(self, arg: ContactGradientMode, /) -> None: ...
-
-    @property
-    def num_dofs(self) -> int: ...
-
-    @property
-    def num_efforts(self) -> int: ...
-
-    @property
-    def state_vector(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @state_vector.setter
-    def state_vector(self, arg: object, /) -> None: ...
-
-    @property
-    def control_vector(self) -> Annotated[NDArray[numpy.float64], dict(shape=(None,), order='C')]: ...
-
-    @control_vector.setter
-    def control_vector(self, arg: object, /) -> None: ...
-
-    def get_step_derivatives(self) -> StepDerivatives: ...
-
-    def apply_step_vjp(self, d_loss_d_next_state: object) -> StepGradient: ...
-
-    def add_differentiable_parameter(self, body: RigidBody, parameter: PhysicalParameter = PhysicalParameter.MASS, *, lower_bound: object | None = ..., upper_bound: object | None = ...) -> None: ...
-
-    @property
-    def num_differentiable_parameters(self) -> int: ...
-
-    def collide(self, options: CollisionQueryOptions = ...) -> list[Contact]: ...
-
-    def clear(self) -> None: ...
-
-    def __repr__(self) -> str: ...
-
-def is_accelerated_deformable_solve_available() -> bool: ...
-
-def set_accelerated_deformable_solve(enabled: bool) -> bool: ...
-
-def is_accelerated_deformable_solve_enabled() -> bool: ...
